requirements.txt
# Core LLM & orchestration
openai>=1.30.0

# Web API framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Rate limiting
slowapi>=0.1.9

# Data validation (used by FastAPI + our models)
pydantic>=2.7.0

# Environment variable management
python-dotenv>=1.0.0

# Development / testing
pytest>=8.2.0
httpx>=0.27.0         # Required for FastAPI TestClient
